Import code from rpmlib.rs
Import (sans history) of the code originally from:

https://github.com/iqlusion-io/crates/tree/master/rpmlib

Some elements deemed incomplete or questionably implemented have been
removed, with the goal of re-adding them with tests.

Note that none of this code is particularly well vetted or tested, so
if you come upon this particular commit as the origin of a line of code,
treat it with caution. It may not be implemented correctly.